eliminate debug api calls
